Rewrite and enhance the application in Python language
In each picture file replace specified color with transparency.

The files in previous commits were written in Bash language.

Use Python version 3.7 .
This application requires python version 3.7 or higher, because the argument 'capture_output' of the method 'subprocess.run' was added in python 3.7 .

Prepare the application for all possible situations and errors.

Enhance the application.

Use optional and positional arguments.
Run this application from command line with argument '--help' or '-h' to get full description of the application and its features.